refactor(desktop): promote AgentIcon and splitFrontMatter to shared homes
Both now have consumers outside their original feature folders (the new
Memory tab), so per the co-location rules they move to the highest shared
parent: AgentIcon to renderer/components, and splitFrontMatter (plus its
test) next to TipTapMarkdownRenderer, whose front-matter limitation it
exists for.
