+++ apps/desktop/src/renderer/components/AgentIcon/AgentIcon.tsx
