+++ apps/desktop/src/renderer/components/AgentIcon/index.ts
